+++ src/main/java/com/gitlab/cdagaming/craftpresence/impl/Tuple.java
@@ -1,4 +1,4 @@
-package com.gitlab.cdagaming.craftpresence.utils;
+package com.gitlab.cdagaming.craftpresence.impl;
 
 public class Tuple<U, V> {