+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/KeyUtils.java
@@ -2,6 +2,7 @@ package com.gitlab.cdagaming.craftpresence.utils;
 
 import com.gitlab.cdagaming.craftpresence.CraftPresence;
 import com.gitlab.cdagaming.craftpresence.ModUtils;
+import com.gitlab.cdagaming.craftpresence.impl.Tuple;
 import com.google.common.collect.Lists;
 import org.lwjgl.input.Keyboard;