+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/SystemUtils.java
@@ -1,7 +1,6 @@
 package com.gitlab.cdagaming.craftpresence.utils;
 
 import com.gitlab.cdagaming.craftpresence.ModUtils;
-import com.gitlab.cdagaming.craftpresence.utils.discord.rpc.DiscordRPC;
 import com.google.common.collect.Lists;
 
 import java.util.List;
@@ -49,7 +48,7 @@ public class SystemUtils {
         // Every Two Seconds, refresh Callbacks
         if (ELAPSED_TIME % 2 == 0) {
             if (!refreshedCallbacks) {
-                DiscordRPC.INSTANCE.Discord_RunCallbacks();
+                //.Discord_RunCallbacks();
                 refreshedCallbacks = true;
             }
         } else {