+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/gui/controls/CheckBoxControl.java
@@ -1,8 +1,8 @@
 package com.gitlab.cdagaming.craftpresence.utils.gui.controls;
 
 import com.gitlab.cdagaming.craftpresence.CraftPresence;
+import com.gitlab.cdagaming.craftpresence.impl.Tuple;
 import com.gitlab.cdagaming.craftpresence.utils.StringUtils;
-import com.gitlab.cdagaming.craftpresence.utils.Tuple;
 import net.minecraft.client.Minecraft;
 
 @SuppressWarnings("NullableProblems")