+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/world/BiomeUtils.java
@@ -1,9 +1,10 @@
 package com.gitlab.cdagaming.craftpresence.utils.world;
 
 import com.gitlab.cdagaming.craftpresence.CraftPresence;
+import com.gitlab.cdagaming.craftpresence.impl.Tuple;
 import com.gitlab.cdagaming.craftpresence.utils.FileUtils;
 import com.gitlab.cdagaming.craftpresence.utils.StringUtils;
-import com.gitlab.cdagaming.craftpresence.utils.Tuple;
+import com.gitlab.cdagaming.craftpresence.utils.discord.rpc.entities.pipe.PipeStatus;
 import com.google.common.collect.Lists;
 import net.minecraft.world.biome.Biome;
 
@@ -32,7 +33,9 @@ public class BiomeUtils {
         isInUse = false;
         CraftPresence.CLIENT.initArgumentData("&BIOME&");
         CraftPresence.CLIENT.initIconData("&BIOME&");
-        CraftPresence.CLIENT.updatePresence(CraftPresence.CLIENT.buildRichPresence());
+        if (CraftPresence.CLIENT.ipcInstance.getStatus() == PipeStatus.CONNECTED) {
+            CraftPresence.CLIENT.updatePresence(CraftPresence.CLIENT.buildRichPresence());
+        }
     }
 
     public void onTick() {